Revert "feat(proof-worker): add proof worker cli with gen-proof command (#5)"
This reverts commit cef711d9bd45235fef63ce1e30473cb8c8ac7f23.
