--- packages/proof-worker/cli.js
@@ -1,6 +1,0 @@
-#!/usr/bin/env node
-
-// Copyright 2024-2025 kzero authors & contributors
-// SPDX-License-Identifier: GNU General Public License v3.0
-
-import './build/index.js';
