+++ packages/proof-worker/package.json
@@ -29,10 +29,8 @@
       }
     }
   },
-  "bin": "./cli.js",
   "files": [
     "build",
-    "cli.js",
     "README.md"
   ],
   "scripts": {
@@ -43,7 +41,6 @@
   "devDependencies": {
     "@types/fs-extra": "^11",
     "@types/ws": "^8",
-    "@types/yargs": "^17.0.33",
     "npm-run-all2": "^7.0.2",
     "typescript": "^5.8.2"
   },
@@ -52,7 +49,6 @@
     "@kzero/logger": "workspace:*",
     "dotenv": "^16.4.7",
     "fs-extra": "^11.3.0",
-    "ws": "^8.18.1",
-    "yargs": "^17.7.2"
+    "ws": "^8.18.1"
   }
 }
