+++ packages/proof-worker/src/index.ts
@@ -3,91 +3,49 @@
 
 import './dotenv.js';
 
-import { randomBytes } from 'node:crypto';
-import yargs from 'yargs';
-import { hideBin } from 'yargs/helpers';
-
-import { generateZKInput, type JWTPublicKeyData } from '@kzero/common';
+import type { SuiProofFields, ZKLoginInput } from '@kzero/common';
 
 import { generateProof } from './generateProof.js';
-import { startWorker } from './worker.js';
+import { WebSocketClient } from './WebSocketClient.js';
 
-yargs(hideBin(process.argv))
-  .scriptName('proof-worker')
-  .usage('$0 <cmd>')
-  .wrap(null)
-  .locale('en')
-  .command('start', 'Start the proof worker', () => {
-    startWorker();
-  })
-  .command(
-    'gen-proof',
-    'Generate a proof',
-    (yargs) =>
-      yargs
-        .option('jwt', {
-          alias: 'j',
-          type: 'string',
-          description: 'jwt to generate proof for'
-        })
-        .option('salt', {
-          alias: 's',
-          type: 'string',
-          description: 'salt to generate proof for'
-        })
-        .option('epoch', {
-          alias: 'e',
-          type: 'string',
-          description: 'epoch for proof'
-        })
-        .option('key', {
-          alias: 'k',
-          type: 'string',
-          description: 'ephemeralPublicKey generate proof for'
-        })
-        .option('randomness', {
-          alias: 'r',
-          type: 'string',
-          description: 'randomness for generate proof'
-        })
-        .option('cert-url', {
-          alias: 'c',
-          type: 'string',
-          description: 'cert url for generate proof, e.g. https://www.googleapis.com/oauth2/v3/certs'
-        })
-        .demandOption(['jwt', 'key', 'epoch', 'randomness', 'cert-url']),
-    async (argv) => {
-      const jwt = argv.jwt;
-      const key = argv.key;
-      const epoch = argv.epoch;
-      const randomness = argv.randomness;
-      const salt = argv.salt || BigInt(randomBytes(32).toString('hex')).toString();
-      const certUrl = argv['cert-url'];
+/**
+ * Message types for WebSocket communication
+ */
+interface WebSocketMessage {
+  task: 'generateProof';
+  proofId: string;
+  payload: {
+    inputs: ZKLoginInput;
+    fields: SuiProofFields;
+  };
+}
 
-      const certs: JWTPublicKeyData[] = await fetch(certUrl)
-        .then((res) => res.json())
-        .then((json: any) => json.keys);
+/**
+ * Starts the proof generation worker
+ * Establishes WebSocket connection and handles proof generation requests
+ */
+export function startWorker(): void {
+  const wsUrl = process.env.PROOF_SERVER_WS_URL;
 
-      generateZKInput({
-        jwt,
-        salt,
-        epoch,
-        keyStr: key as `0x${string}`,
-        randomness,
-        certs
-      })
-        .then(({ inputs, fields }) => {
-          return generateProof(inputs, fields);
+  if (!wsUrl) {
+    throw new Error('PROOF_SERVER_WS_URL environment variable is not set');
+  }
+
+  const ws = new WebSocketClient(wsUrl);
+
+  ws.on('message', async (message: Buffer) => {
+    const data = JSON.parse(message.toString('utf-8')) as WebSocketMessage;
+
+    if (data.task === 'generateProof') {
+      const result = await generateProof(data.payload.inputs, data.payload.fields);
+
+      ws.send(
+        JSON.stringify({
+          task: 'generateProof',
+          proofId: data.proofId,
+          results: result
         })
-        .then(() => {
-          process.exit();
-        })
-        .catch(() => {
-          process.exit(1);
-        });
+      );
     }
-  )
-  .demandCommand(1, 'You must provide a valid command.')
-  .help()
-  .version(false)
-  .strict().argv;
+  });
+}
